+++ Media Blocks SDK/WPF/CSharp/RTSP Preview Demo/Window1.xaml.cs
@@ -1,0 +1,303 @@
+﻿// ReSharper disable InconsistentNaming
+
+// ReSharper disable StyleCop.SA1600
+// ReSharper disable StyleCop.SA1601
+
+namespace RTSP_Preview
+{
+    using System;
+    using System.Diagnostics;
+    using System.IO;
+    using System.Timers;
+    using System.Windows;
+    using System.Windows.Input;
+
+    using VisioForge.Core;
+    using VisioForge.Core.MediaBlocks;
+    using VisioForge.Core.MediaBlocks.AudioRendering;
+    using VisioForge.Core.MediaBlocks.Sources;
+    using VisioForge.Core.MediaBlocks.VideoRendering;
+    using VisioForge.Core.ONVIF.Legacy;
+    using VisioForge.Core.Types.Events;
+    using VisioForge.Core.Types.X.Sources;
+    using VisioForge.Core.UI;
+    using VisioForge.Core.VideoCaptureX;
+    using Application = System.Windows.Forms.Application;
+    using Task = System.Threading.Tasks.Task;
+    using Uri = System.Uri;
+
+    public partial class Window1 : IDisposable
+    {
+        private Timer tmRecording = new Timer(1000);
+
+        private ONVIFControl onvifControl;
+
+        private MediaBlocksPipeline _pipeline;
+
+        private RTSPSourceBlock _rtspSource;
+
+        private MediaBlock _videoRenderer;
+
+        private AudioRendererBlock _audioRenderer;
+
+        private bool disposedValue;
+
+        public Window1()
+        {
+            InitializeComponent();
+
+            Application.EnableVisualStyles();
+
+            // We have to initialize the engine on start
+            VisioForgeX.InitSDK();
+        }
+
+        private void CreateEngine()
+        {
+            _pipeline = new MediaBlocksPipeline(true);
+
+            _pipeline.OnError += VideoCapture1_OnError;
+        }
+
+        private async Task DestroyEngineAsync()
+        {
+            if (_pipeline != null)
+            {
+                _pipeline.OnError -= VideoCapture1_OnError;
+
+                await _pipeline.DisposeAsync();
+                _pipeline = null;
+            }
+        }
+
+        private void Form1_Load(object sender, RoutedEventArgs e)
+        {
+            Title += $" (SDK v{VideoCaptureCoreX.SDK_Version})";
+
+            tmRecording.Elapsed += async (senderx, args) =>
+            {
+                await UpdateRecordingTime();
+            };
+        }
+
+        private async void btStart_Click(object sender, RoutedEventArgs e)
+        {
+            CreateEngine();
+
+            if (onvifControl != null)
+            {
+                onvifControl.Disconnect();
+                onvifControl.Dispose();
+                onvifControl = null;
+
+                btONVIFConnect.Content = "Connect";
+            }
+
+            mmLog.Clear();
+
+            var audioEnabled = cbIPAudioCapture.IsChecked == true; 
+            _pipeline.Debug_Mode = cbDebugMode.IsChecked == true;
+            _pipeline.Debug_Dir = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments), "VisioForge");
+
+            var rtsp = await RTSPSourceSettings.CreateAsync(new Uri(cbIPURL.Text), edIPLogin.Text, edIPPassword.Text, audioEnabled);
+
+            _rtspSource = new RTSPSourceBlock(rtsp);
+
+            _videoRenderer = new VideoRendererBlock(_pipeline, VideoView1);
+
+            _pipeline.Connect(_rtspSource.VideoOutput, _videoRenderer.Input);
+
+            if (audioEnabled && rtsp.GetInfo().AudioStreams.Count > 0)
+            {
+                _audioRenderer = new AudioRendererBlock();
+                _pipeline.Connect(_rtspSource.AudioOutput, _audioRenderer.Input);
+            }
+
+            await _pipeline.StartAsync();
+
+            tmRecording.Start();
+        }
+
+        private async void btStop_Click(object sender, RoutedEventArgs e)
+        {
+            tmRecording.Stop();
+
+            await _pipeline.StopAsync();
+            await _pipeline.DisposeAsync();
+        }
+
+        private void llVideoTutorials_LinkClicked(object sender, MouseButtonEventArgs e)
+        {
+            var startInfo = new ProcessStartInfo("explorer.exe", HelpLinks.VideoTutorials);
+            Process.Start(startInfo);
+        }
+
+        private void Log(string txt)
+        {
+            Dispatcher.Invoke(() => { mmLog.Text = mmLog.Text + txt + Environment.NewLine; });
+        }
+
+        private void VideoCapture1_OnError(object sender, ErrorsEventArgs e)
+        {
+            Log(e.Message);
+        }
+
+        private async void btONVIFConnect_Click(object sender, RoutedEventArgs e)
+        {
+            if (btONVIFConnect.Content.ToString() == "Connect")
+            {
+                try
+                {
+                    btONVIFConnect.IsEnabled = false;
+                    btONVIFConnect.Content = "Connecting";
+
+                    if (onvifControl != null)
+                    {
+                        onvifControl.Disconnect();
+                        onvifControl.Dispose();
+                        onvifControl = null;
+                    }
+
+                    if (string.IsNullOrEmpty(edONVIFLogin.Text) || string.IsNullOrEmpty(edONVIFPassword.Text))
+                    {
+                        MessageBox.Show(this, "Please specify IP camera user name and password.");
+                        return;
+                    }
+
+                    onvifControl = new ONVIFControl();
+                    var result = await onvifControl.ConnectAsync(edONVIFURL.Text, edONVIFLogin.Text, edONVIFPassword.Text);
+                    if (!result)
+                    {
+                        onvifControl = null;
+                        MessageBox.Show(this, "Unable to connect to ONVIF camera.");
+                        return;
+                    }
+
+                    var deviceInfo = await onvifControl.GetDeviceInformationAsync();
+                    if (deviceInfo != null)
+                    {
+                        lbONVIFCameraInfo.Content = $"Model {deviceInfo.Model}, Firmware {deviceInfo.Firmware}";
+                    }
+
+                    cbONVIFProfile.Items.Clear();
+                    var profiles = await onvifControl.GetProfilesAsync();
+                    foreach (var profile in profiles)
+                    {
+                        cbONVIFProfile.Items.Add($"{profile.Name}");
+                    }
+
+                    if (cbONVIFProfile.Items.Count > 0)
+                    {
+                        cbONVIFProfile.SelectedIndex = 0;
+                    }
+
+                    edONVIFLiveVideoURL.Text = cbIPURL.Text = await onvifControl.GetVideoURLAsync();
+
+                    edIPLogin.Text = edONVIFLogin.Text;
+                    edIPPassword.Text = edONVIFPassword.Text;
+
+                    btONVIFConnect.Content = "Disconnect";
+                }
+                finally
+                {
+                    btONVIFConnect.IsEnabled = true;
+                    btONVIFConnect.Content = "Connect";
+                }
+            }
+            else
+            {
+                btONVIFConnect.Content = "Connect";
+
+                if (onvifControl != null)
+                {
+                    onvifControl.Disconnect();
+                    onvifControl.Dispose();
+                    onvifControl = null;
+                }
+            }
+        }
+
+        private async void btResume_Click(object sender, RoutedEventArgs e)
+        {
+            await _pipeline.ResumeAsync();
+        }
+
+        private async void btPause_Click(object sender, RoutedEventArgs e)
+        {
+            await _pipeline.PauseAsync();
+        }
+
+        private async Task UpdateRecordingTime()
+        {
+            var ts = await _pipeline.Position_GetAsync();
+
+            if (Math.Abs(ts.TotalMilliseconds) < 0.01)
+            {
+                return;
+            }
+
+            Dispatcher.Invoke((Action)(() =>
+            {
+                lbTimestamp.Text = "Duration: " + ts.ToString(@"hh\:mm\:ss");
+            }));
+        }
+
+        private async void btListONVIFSources_Click(object sender, RoutedEventArgs e)
+        {
+            cbIPURL.Items.Clear();
+
+            var lst = await DeviceEnumerator.Shared.ONVIF_ListSourcesAsync(null, null);
+            foreach (var uri in lst)
+            {
+                cbIPURL.Items.Add(uri);
+            }
+
+            if (cbIPURL.Items.Count > 0)
+            {
+                cbIPURL.SelectedIndex = 0;
+            }
+        }
+
+        private async void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e)
+        {
+            await DestroyEngineAsync();
+
+            VisioForgeX.DestroySDK();
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (!disposedValue)
+            {
+                if (disposing)
+                {
+                    if (onvifControl != null)
+                    {
+                        onvifControl.Dispose();
+                        onvifControl = null;
+                    }
+
+                    tmRecording?.Dispose();
+                    tmRecording = null;
+                }
+
+                disposedValue = true;
+            }
+        }
+
+        ~Window1()
+        {
+            // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
+            Dispose(disposing: false);
+        }
+
+        public void Dispose()
+        {
+            // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
+            Dispose(disposing: true);
+            GC.SuppressFinalize(this);
+        }
+    }
+}
+
+// ReSharper restore InconsistentNaming